feat(215): LinkML schema + generation for Storyboard/Scene/Viewport
- Extend FeatureKindEnum with STORYBOARD and STORYBOARD_SCENE
- Add optional agent slot to LogEntry (Article III single provenance surface)
- Create storyboard.yaml defining StoryboardProperties, SceneProperties, Viewport,
  StoryboardFeature, SceneFeature with reserved slots (time_range null, bearing=0)
- Regenerate Pydantic, JSON Schema, and TypeScript bindings
- Add immer ^10.1.3 and ulid ^3.0.2 to shared/components deps
- Scaffold shared/components/src/storyboard/ module (barrel only)
- Lift sha256Hex into shared/components/src/utils/hash.ts; re-export from nl-cql2/hash.ts
