+++ shared/components/package.json
@@ -99,7 +99,9 @@
     "vega-embed": "^6.26.0",
     "vega-lite": "^5.21.0",
     "vscrui": "^0.1.0",
-    "modern-screenshot": "^4.5.0"
+    "modern-screenshot": "^4.5.0",
+    "immer": "^10.1.3",
+    "ulid": "^3.0.2"
   },
   "devDependencies": {
     "@playwright/test": "^1.58.0",
