+++ shared/components/src/nl-cql2/hash.ts
@@ -1,29 +1,13 @@
 /**
  * SHA-256 hashing utility for the NL → CQL2 generator (#188).
  *
- * Uses the Web Crypto API (`crypto.subtle.digest`), which is available in
- * modern Node (18+) and all evergreen browsers — keeps the module
- * browser-compatible per the plan's Target Platform requirement.
+ * `sha256Hex` has been lifted to `shared/components/src/utils/hash.ts` in
+ * feature 215 so the storyboard CRUD module can reuse the same primitive.
+ * This module re-exports the canonical helper, and also owns the nl-cql2-
+ * specific `canonicalisePhrase` helper.
  */
 
-/** Return the SHA-256 of `input` as a lowercase hex string. */
-export async function sha256Hex(input: string): Promise<string> {
-  const data = new TextEncoder().encode(input);
-  const subtle = globalThis.crypto?.subtle;
-  if (!subtle) {
-    throw new Error(
-      "[nl-cql2/hash] globalThis.crypto.subtle is not available; " +
-        "run under Node 18+ or a modern browser.",
-    );
-  }
-  const digest = await subtle.digest("SHA-256", data);
-  const bytes = new Uint8Array(digest);
-  let hex = "";
-  for (const b of bytes) {
-    hex += b.toString(16).padStart(2, "0");
-  }
-  return hex;
-}
+export { sha256Hex } from "../utils/hash";
 
 /**
  * Canonicalise an analyst phrase for fixture lookup: trim, lowercase, collapse
