+++ shared/components/src/storyboard/index.ts
@@ -1,0 +1,3 @@
+// Public barrel for the storyboard CRUD module (feature 215).
+// Populated by downstream tasks T040+.
+export {};
